Sidebar arrows should change content when open
The sidebar arrows should change content when they're open.  There's a small bug in css/customstyles.css that prevents this; other parts of the style use 'active' instead of 'open' so it needs to here too.

This PR fixes it.

Since Font Awesome is already wired in, I think that their carets look nicer.  Here's how:

```
.nav li > a > span:after {
   font-family: FontAwesome;
   content: "\f105";

}
.nav li.active > a > span:after {
   font-family: FontAwesome;
   content: "\f107";
}
```